requirements.txt
# Python зависимости для автоматизации Tiled
# Установка: pip3 install -r requirements.txt

# Работа с изображениями
Pillow>=10.0.0
